new_scaling_bloom() does not need an id, 0 always works
The id argument is the starting id of the first counting bloom in
the scaling bloom. It might as well be the lowest possible id,
because for an item with a "very low" id, the user definitely wants
to put it at least in the first counting bloom anyway, instead of
rejecting it.

The starting id of the next counting bloom is set to the highest id
of any item when the first counting bloom fills, and thus is not
really affected by the initial id.