new default and options for determining PY_MOD_DIR for pydablooms
Now defaults to a python path in /usr/local if it exists, and one
can specify PY_MOD_DIR_ARG=--user to try to find the python path
in HOME which `pip install --user` would use.

This may be more convenient for OS X 10.11 El Capitan since it
does not allow even root to write to "system" dirs.