Make Destroy public
I think this method needs to be public in order for the backing files to be closed. According to lsof on my system, simply freeing the Go object leaves the file descriptor open forever. I happen to always know when my filter is going to go out of scope, so I can call Destroy and make sure the fds get freed.